A second version of parse-svd that generates only single-value tables
They form a tree, but each node can only have one index (the "key") and
its value can either be a piece of text, or a subtree.

This looks very nice when printed out, and the *order* of the nodes of
the resulting tree is identical to that of the parsed XML, but it's less
useful for *processing*.

I want to make a version that merges all the single-value tables that
exist at the same level into one table.