Adding a list of BROKEN chips; these do not get processed
When doing a

  make everything

I found a few files that barfed, mostly because their .svd files had
weird stuff that confused my parser, or had a multiple-register
definition (using "dim =") that was missing a "dim-index" entry. Instead
of wasting time with these files, I've added them to a BROKEN list in
the Makefile. They are not processed.

I've also removed the dependency on the Makefile for the .lua and .mu4
files. It is annoying. The important thing is the the .lua files get
re-made when the parser (parse-svd.lua) changes, and the .mu4 files get
re-made when the register printer (print-regs.lua) changes.