print-regs-generic.lua can create register "clusters" (<dim>)
The <dim> element in the SVD file - it seems someone was spending too
much time programming in BASIC or FORTRAN - allows a single register
declaration to create a series of registers whose names follow
a template with a placeholder - "%s" - that is replaced with a value for
the specific register. There is a memory offset between the instances
- dimIncrement - and dimIndex can be used to name them. Normally they
start at "0", but dimIndex can contain, eg, "4-24".

Overall it's a stupid idea. The XML file might as well be as simple as
possible, without any templating powers. It should simply list all the
registers.

But since the ESP32 SVD use this feature, I had to add support for it.